+++ src/main/java/ar/edu/utn/frba/dds/repositories/RepoDeMiembros.java
@@ -41,7 +41,7 @@ public class RepoDeMiembros  implements WithSimplePersistenceUnit {
         entityManager().merge(miembro);
         tx.commit();
     }
-    public Miembro buscarPorId(Integer id){
+    public Miembro buscarPorId(Long id){
         return entityManager().find(Miembro.class,id);
     }
     public List<Miembro> buscarTodos(){
